@@ -1,17 +1,23 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import './index.css'
+import App from './App'
+import {BrowserRouter} from "react-router-dom";
+import store from './redux/store';
+import { Provider } from 'react-redux';
 
-ReactDOM.render(
+
+
+const root = ReactDOM.createRoot(document.getElementById("root"))
+
+root.render(
+    <Provider store={store}>
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+  <BrowserRouter>
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+    <App/>
+ 
+  </BrowserRouter>
+    </React.StrictMode>
+  </Provider>
+)
